LayoutTests/media/video-seek-after-end-play-expected.txt

Test that calling play while a seek is ongoing will resume.


EVENT(play)
EVENT(playing)
EVENT(ended)
EVENT(seeking)
EVENT(seeked)
EVENT(play)
EVENT(seeking)
EVENT(seeked)
EVENT(playing)
EVENT(ended)
END OF TEST
